Revert "[Opt](compile) Fix some compile checks (#53318)"
This reverts commit bfb52e20448b4f1a9171b50c10e92239166c29c4.

+++ be/src/olap/primary_key_index.h
@@ -34,7 +34,6 @@
 #include "util/slice.h"
 
 namespace doris {
-#include "common/compile_check_begin.h"
 class TypeInfo;
 
 namespace io {
@@ -133,7 +132,7 @@ public:
         return _bf->test_bytes(key.data, key.size);
     }
 
-    int64_t num_rows() const {
+    uint32_t num_rows() const {
         DCHECK(_index_parsed);
         return _index_reader->num_values();
     }
@@ -158,5 +157,5 @@ private:
     size_t _bf_num = 0;
     uint64_t _bf_bytes = 0;
 };
-#include "common/compile_check_end.h"
+
 } // namespace doris
